refactor: Move Polyfill to it's own namespace
