+++ lib/Libs/Polyfill/Php80/Resources/Stubs/Stringable.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Resque\Libs\Polyfill\Php80\Resources\Stubs;
+
+if (\PHP_VERSION_ID < 80000) {
+    interface Stringable
+    {
+        /**
+         * @return string
+         */
+        public function __toString();
+    }
+}
